+++ tests/testthat/test-word.r
@@ -6,3 +6,7 @@ test_that("word extraction", {
   expect_equal("moon", word("walk the moon", 3))
   expect_equal("the moon", word("walk the moon", 2, 3))
 })
+
+test_that("words past end return NA", {
+  expect_equal(word("a b c", 4), NA_character_)
+})